pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gitbit"
version = "0.1.0"
description = "A comprehensive Git repository management and version control operations agent"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
dependencies = [
    "gitpython>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "mypy>=0.990",
    "pylint>=2.15",
]

[tool.black]
line-length = 100
target-version = ["py37"]
include = '\.pyi?$'

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"